Split worker crates out of waymark
