+++ Cargo.toml
@@ -37,6 +37,12 @@ waymark-pool-status = { path = "crates/lib/pool-status" }
 waymark-webapp-config = { path = "crates/lib/webapp-config" }
 waymark-webapp-routes = { path = "crates/lib/webapp-routes" }
 waymark-webapp-bringup = { path = "crates/lib/webapp-bringup" }
+waymark-worker-core = { path = "crates/lib/worker-core" }
+waymark-worker-remote = { path = "crates/lib/worker-remote" }
+waymark-worker-inline = { path = "crates/lib/worker-inline" }
+waymark-worker-status-core = { path = "crates/lib/worker-status-core" }
+waymark-worker-status-reporter = { path = "crates/lib/worker-status-reporter" }
+waymark-worker-metrics = { path = "crates/lib/worker-metrics" }
 
 anyhow = "1"
 async-trait = "0.1"
@@ -45,7 +51,7 @@ chrono = { version = "0.4", default-features = false }
 clap = "4.5"
 console-subscriber = "0.5"
 cron = "0.12"
-futures = "0.3"
+futures-core = "0.3"
 http-body-util = "0.1"
 proptest = "1.9"
 prost = "0.12"
