+++ crates/bin/integration-test/Cargo.toml
@@ -14,6 +14,7 @@ waymark-backend-postgres-migrations = { workspace = true }
 waymark-workflow-registry-backend = { workspace = true }
 waymark-backend-memory = { workspace = true }
 waymark-proto = { workspace = true }
+waymark-worker-remote = { workspace = true }
 waymark = { workspace = true }
 
 anyhow = { workspace = true }
