+++ crates/bin/integration-test/src/main.rs
@@ -19,7 +19,6 @@ use sqlx::Row;
 use uuid::Uuid;
 
 use waymark::waymark_core::runloop::{RunLoop, RunLoopConfig};
-use waymark::workers::{PythonWorkerConfig, RemoteWorkerPool};
 use waymark_backend_memory::MemoryBackend;
 use waymark_backend_postgres::PostgresBackend;
 use waymark_core_backend::{CoreBackend, QueuedInstance};
@@ -28,6 +27,7 @@ use waymark_ir_conversions::literal_from_json_value;
 use waymark_proto::ast as ir;
 use waymark_runner_state::RunnerState;
 use waymark_support_integration::{LOCAL_POSTGRES_DSN, connect_pool, ensure_local_postgres};
+use waymark_worker_remote::{PythonWorkerConfig, RemoteWorkerPool};
 use waymark_workflow_registry_backend::{WorkflowRegistration, WorkflowRegistryBackend};
 
 #[derive(Parser, Debug)]
