+++ crates/bin/smoke/src/main.rs
@@ -18,7 +18,6 @@ use waymark_workflow_registry_backend::{WorkflowRegistration, WorkflowRegistryBa
 use waymark::waymark_core::dag_viz::render_dag_image;
 use waymark::waymark_core::ir_format::format_program;
 use waymark::waymark_core::runloop::{RunLoop, RunLoopConfig};
-use waymark::workers::{PythonWorkerConfig, RemoteWorkerPool};
 use waymark_dag::convert_to_dag;
 use waymark_ir_conversions::literal_from_json_value;
 use waymark_proto::ast as ir;
@@ -27,6 +26,7 @@ use waymark_smoke_sources::{
     build_control_flow_program, build_parallel_spread_program, build_program,
     build_try_except_program, build_while_loop_program,
 };
+use waymark_worker_remote::{PythonWorkerConfig, RemoteWorkerPool};
 
 #[derive(Parser, Debug)]
 #[command(name = "waymark-smoke", about = "Smoke check core-python components.")]
