+++ crates/lib/worker-inline/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "waymark-worker-inline"
+edition = "2024"
+version.workspace = true
+
+[dependencies]
+serde_json = { workspace = true }
+tokio = { workspace = true }
+waymark-observability = { workspace = true }
+waymark-worker-core = { workspace = true }
