+++ crates/lib/worker-remote/Cargo.toml
@@ -1,0 +1,21 @@
+[package]
+name = "waymark-worker-remote"
+edition = "2024"
+version.workspace = true
+
+[dependencies]
+anyhow = { workspace = true } # TODO: drop
+futures-core = { workspace = true }
+prost = { workspace = true }
+serde_json = { workspace = true }
+thiserror = { workspace = true }
+tokio = { workspace = true }
+tokio-stream = { workspace = true, features = ["net"] }
+tonic = { workspace = true }
+tracing = { workspace = true }
+uuid = { workspace = true }
+waymark-message-conversions = { workspace = true }
+waymark-proto = { workspace = true }
+waymark-worker-core = { workspace = true }
+waymark-worker-metrics = { workspace = true }
+waymark-worker-status-core = { workspace = true }
