+++ crates/lib/worker-status-backend/Cargo.toml
@@ -5,6 +5,6 @@ edition = "2024"
 
 [dependencies]
 async-trait = { workspace = true }
-uuid = { workspace = true }
 chrono = { workspace = true }
+uuid = { workspace = true }
 waymark-backends-core = { workspace = true }
