+++ crates/lib/worker-status-core/Cargo.toml
@@ -1,0 +1,7 @@
+[package]
+name = "waymark-worker-status-core"
+edition = "2024"
+version.workspace = true
+
+[dependencies]
+chrono = { workspace = true }
