+++ crates/lib/worker-status-reporter/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "waymark-worker-status-reporter"
+edition = "2024"
+version.workspace = true
+
+[dependencies]
+chrono = { workspace = true }
+tokio = { workspace = true }
+tokio-util = { workspace = true }
+tracing = { workspace = true }
+uuid = { workspace = true }
+waymark-pool-status = { workspace = true }
+waymark-worker-status-backend = { workspace = true }
+waymark-worker-status-core = { workspace = true }
