+++ crates/waymark/src/lib.rs
@@ -2,18 +2,9 @@
 
 pub mod config;
 pub mod garbage_collector;
-pub mod messages;
 pub mod scheduler;
-pub mod server_worker;
 pub mod waymark_core;
-pub mod workers;
 
 // Worker infrastructure (preserved from the legacy Rust core).
 pub use garbage_collector::{GarbageCollectorConfig, GarbageCollectorTask};
-pub use messages::MessageError;
 pub use scheduler::{SchedulerConfig, SchedulerTask};
-pub use server_worker::{WorkerBridgeChannels, WorkerBridgeServer};
-pub use workers::{
-    ActionDispatchPayload, PythonWorker, PythonWorkerConfig, PythonWorkerPool, RemoteWorkerPool,
-    RoundTripMetrics, spawn_status_reporter,
-};
