+++ crates/waymark/src/waymark_core/mod.rs
@@ -6,7 +6,6 @@ pub mod ir_format;
 pub mod lock;
 pub mod runloop;
 
-pub use crate::workers::{ActionCompletion, ActionRequest, BaseWorkerPool, InlineWorkerPool};
 pub use dag_viz::{build_dag_graph, render_dag_image};
 pub use ir_format::format_program;
 pub use runloop::RunLoop;
