--- crates/waymark/src/workers/mod.rs
@@ -1,16 +1,0 @@
-//! Worker pool implementations.
-
-mod base;
-mod inline;
-mod remote;
-mod status;
-
-pub use base::{
-    ActionCompletion, ActionRequest, BaseWorkerPool, RoundTripMetrics, WorkerPoolError,
-    WorkerThroughputSnapshot,
-};
-pub use inline::{ActionCallable, InlineWorkerPool};
-pub use remote::{
-    ActionDispatchPayload, PythonWorker, PythonWorkerConfig, PythonWorkerPool, RemoteWorkerPool,
-};
-pub use status::{WorkerPoolStats, WorkerPoolStatsSnapshot, spawn_status_reporter};
